requirements.txt
# This requirements file lists all third-party dependencies for this project.
#
# Run 'pip install -r requirements.txt -t lib/' to install these dependencies
# in `lib/` subdirectory.
#
# Note: The `lib` directory is added to `sys.path` by `appengine_config.py`.
flask
flask_login
beautifulsoup4
bleach
oauth2client
google-api-python-client
python-amazon-simple-product-api
